turret_controller/requirements.txt
pyserial>=3.5
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0

